+ Fixed using big estimates (>600000h). Should work now with very big numbers
git-svn-id: http://agilefant.googlecode.com/svn/trunk@1292 29784df9-99d5-7e28-3e6c-ac1830307ffc

+++ src/fi/hut/soberit/agilefant/db/hibernate/TimeUserType.java
@@ -19,7 +19,7 @@ import fi.hut.soberit.agilefant.model.AFTime;
  */
 public class TimeUserType implements UserType {
 
-    private static final int[] SQL_TYPES = { Types.INTEGER };
+    private static final int[] SQL_TYPES = { Types.BIGINT };
 
     /**
      * Get the sql types to use to save our "Time" object.
